+++ src/Controller/ControlPanel/Edit.php
@@ -3,16 +3,16 @@
 namespace Message\Mothership\CMS\Controller\ControlPanel;
 
 use Message\Cog\Field;
+use Message\Cog\DB\NestedSetException;
 
 use Message\Mothership\CMS\Page\Authorisation;
 use Message\Mothership\CMS\Page\Page;
 use Message\Mothership\CMS\Page\Content;
-use Message\Cog\Field\Form;
-use Message\Cog\Field\Factory;
-use Message\Cog\Field\RepeatableContainer;
+use Message\Mothership\CMS\Page\Exception\PageEditException;
 
 use Message\Cog\ValueObject\Slug;
 use Message\Mothership\FileManager\File;
+use Message\Mothership\CMS\Page\Exception\InvalidSlugException;
 
 class Edit extends \Message\Cog\Controller\Controller
 {
@@ -22,6 +22,8 @@ class Edit extends \Message\Cog\Controller\Controller
 	 * Index for editing, this just redirects to the content edit screen.
 	 *
 	 * @param int $pageID The page ID
+	 *
+	 * @return Response
 	 */
 	public function index($pageID)
 	{
@@ -68,6 +70,8 @@ class Edit extends \Message\Cog\Controller\Controller
 	 * POST action for updating the page's title.
 	 *
 	 * @param int $pageID The page ID
+	 *
+	 * @return Response
 	 */
 	public function updateTitle($pageID)
 	{
@@ -93,6 +97,8 @@ class Edit extends \Message\Cog\Controller\Controller
 	 * Render the content form.
 	 *
 	 * @param int $pageID The page ID
+	 *
+	 * @return Response
 	 */
 	public function content($pageID, $form = null)
 	{
@@ -139,6 +145,8 @@ class Edit extends \Message\Cog\Controller\Controller
 	 * Render the attributes form.
 	 *
 	 * @param int $pageID The page ID
+	 *
+	 * @return Response
 	 */
 	public function attributes($pageID)
 	{
@@ -166,6 +174,8 @@ class Edit extends \Message\Cog\Controller\Controller
 	 * @param  int 		$pageID 	The pageID of the page
 	 * @param  string 	$slug   	The slug to remove from history and update
 	 *                          	the given page
+	 *
+	 * @return Response
 	 */
 	public function forceSlugAction($pageID, $slug)
 	{
@@ -191,6 +201,8 @@ class Edit extends \Message\Cog\Controller\Controller
 	 *        	already exist in that new section.
 	 *
 	 * @param  	int 	$pageID 	id of the Page object to be loaded and updated
+	 *
+	 * @return Response
 	 */
 	public function attributesAction($pageID)
 	{
@@ -202,39 +214,46 @@ class Edit extends \Message\Cog\Controller\Controller
 			$data['parent'] = isset($data['parent']) ? $data['parent'] : 0;
 			// If the parentID != the submitted parent OR the parent is false
 			// (as it's root) and the submitted parent is not 0 (root)
-			if (($parent && $parent->id != $data['parent']) || (!$parent && $data['parent'] != 0)) {
-				if ($this->get('cms.page.edit')->changeParent($pageID, $data['parent'])) {
-					$this->addFlash('success', 'Parent successully changed');
-				} else {
-					$this->addFlash('error', 'The page could not be moved to a new position');
+
+			try {
+				if (($parent && $parent->id != $data['parent']) || (!$parent && $data['parent'] != 0)) {
+					$this->get('cms.page.edit')->changeParent($pageID, $data['parent']);
+					$this->addFlash('success', $this->trans('ms.cms.feedback.edit.attributes.parent.success'));
 				}
+
+				if (!is_null($data['siblings']) && $data['siblings'] >= 0) {
+					$index = $data['siblings'];
+					$this->get('cms.page.edit')->changeOrder($page, $index);
+				}
+			} catch (NestedSetException $e) {
+				$this->addFlash('error', $this->trans('ms.cms.feedback.edit.attributes.nested-set.error', ['%error%' => $e->getMessage()]));
+
+				return $this->redirectToReferer();
+			} catch (PageEditException $e) {
+				$this->addFlash('error', $this->trans('ms.cms.feedback.edit.attributes.order.failure'));
+			}
+			if (!$page->isHomepage()) {
+				$page = $this->_updateSlug($page, $data['slug']);
 			}
 
-			if (!is_null($data['siblings']) && $data['siblings'] >= 0) {
-				$index = $data['siblings'];
-				if (!$this->get('cms.page.edit')->changeOrder($page, $index)) {
-					$this->addFlash('error', 'The page could not be moved to a new position');
-				}
-			}
-
-			$page = $this->_updateSlug($page, $data['slug']);
+			$userGroups = $this->get('user.groups');
 
 			$page->visibilitySearch     = $data['visibility_search'];
 			$page->visibilityMenu       = $data['visibility_menu'];
 			$page->visibilityAggregator = $data['visibility_aggregator'];
 			$page->access               = $data['access'] ?: 0;
-			$page->accessGroups         = $data['access_groups'];
+			
+			$page->accessGroups = array_map(function($group) use ($userGroups) {
+				return $userGroups->get($group);
+			}, $data['access_groups']);
+
 			$page->setTags($this->_parseTags($data['tags']));
 
 			$page = $this->get('cms.page.edit')->save($page);
 			$this->addFlash('success', $this->trans('ms.cms.feedback.edit.attributes.success'));
 		}
 
-
-		return $this->render('::edit/attributes', array(
-			'page' => $page,
-			'form' => $form,
-		));
+		return $this->redirectToReferer();
 	}
 
 	/**
@@ -372,7 +391,7 @@ class Edit extends \Message\Cog\Controller\Controller
 		else {
 			$form->add('slug', 'text', $this->trans('ms.cms.attributes.slug.label'), array(
 				'read_only' => true,
-				'data' => $this->trans('ms.cms.attributes.slug.homepage')
+				'data' => $this->trans('ms.cms.attributes.slug.homepage'),
 			));
 		}
 
@@ -412,6 +431,7 @@ class Edit extends \Message\Cog\Controller\Controller
 		))->val()->optional();
 
 		$siblings = $this->get('cms.page.loader')->getSiblings($page);
+
 		$siblingChoices = array();
 		if ($siblings) {
 			$siblingChoices[0] = 'Move to top';
@@ -546,8 +566,21 @@ class Edit extends \Message\Cog\Controller\Controller
 		$slug = '/'.implode('/',$slugSegments);
 		$checkSlug = $this->get('cms.page.loader')->getBySlug($slug, false);
 
+		try {
+			$routes = $this->get('routing.matcher')->match($slug);
+
+			// continue if the frontend route is the most promenant
+			if ($routes['_route'] !== 'ms.cms.frontend') {
+				$this->addFlash('error',  $this->trans('ms.cms.feedback.force-slug.failure.reserved-route'));
+				$update = false;
+			}
+		} catch (\Symfony\Component\Routing\Exception\ResourceNotFoundException $e) {	
+			$this->addFlash('error', $this->trans('ms.cms.feedback.force-slug.failure.not-matched'));
+			$update = false;
+		}
+
 		// If not slug has been found, we need to check the history too
-		if (!$checkSlug) {
+		if (!$checkSlug && $update) {
 			// Check for the slug historicaly and show deleted ones too
 			$historicalSlug = $this->get('cms.page.loader')
 				->includeDeleted(true)
@@ -607,7 +640,13 @@ class Edit extends \Message\Cog\Controller\Controller
 		// If the slug has changed then update the slug
 		if ($update && $page->slug->getLastSegment() != $newSlug) {
 			$this->get('cms.page.edit')->removeHistoricalSlug($slug);
-			$page = $this->get('cms.page.edit')->updateSlug($page, $newSlug);
+			try {
+				$page = $this->get('cms.page.edit')->updateSlug($page, $newSlug);
+			} catch (InvalidSlugException $e) {
+				$this->addFlash('error', $this->trans('ms.cms.feedback.force-slug.failure.generic', [
+					'%message%' => $e->getMessage(),
+				]));
+			}
 		}
 
 		// return the updated or unchanged page